+++ packages/iov-ethereum/src/abi.spec.ts
@@ -54,6 +54,97 @@ describe("Abi", () => {
     });
   });
 
+  describe("encodeUint256", () => {
+    it("works", () => {
+      expect(Abi.encodeUint256("0")).toEqual(
+        fromHex("0000000000000000000000000000000000000000000000000000000000000000"),
+      );
+      expect(Abi.encodeUint256("1")).toEqual(
+        fromHex("0000000000000000000000000000000000000000000000000000000000000001"),
+      );
+      expect(Abi.encodeUint256("2")).toEqual(
+        fromHex("0000000000000000000000000000000000000000000000000000000000000002"),
+      );
+      expect(Abi.encodeUint256("123456789")).toEqual(
+        fromHex("00000000000000000000000000000000000000000000000000000000075bcd15"),
+      );
+      // 2^255
+      expect(
+        Abi.encodeUint256("57896044618658097711785492504343953926634992332820282019728792003956564819968"),
+      ).toEqual(fromHex("8000000000000000000000000000000000000000000000000000000000000000"));
+      // 2^256-1
+      expect(
+        Abi.encodeUint256("115792089237316195423570985008687907853269984665640564039457584007913129639935"),
+      ).toEqual(fromHex("ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff"));
+    });
+
+    it("throws for invalid input", () => {
+      expect(() => Abi.encodeUint256("")).toThrow();
+      expect(() => Abi.encodeUint256(" 1")).toThrow();
+      expect(() => Abi.encodeUint256("-1")).toThrow();
+      expect(() => Abi.encodeUint256("0x1")).toThrow();
+      expect(() => Abi.encodeUint256("a")).toThrow();
+
+      // too large (2^256)
+      expect(() =>
+        Abi.encodeUint256("115792089237316195423570985008687907853269984665640564039457584007913129639936"),
+      ).toThrow();
+    });
+  });
+
+  describe("decodeAddress", () => {
+    it("works", () => {
+      expect(
+        Abi.decodeAddress(fromHex("0000000000000000000000000000000000000000000000000000000000000000")),
+      ).toEqual("0x0000000000000000000000000000000000000000");
+      expect(
+        Abi.decodeAddress(fromHex("0000000000000000000000000000000000000000000000000000000000000001")),
+      ).toEqual("0x0000000000000000000000000000000000000001");
+      expect(
+        Abi.decodeAddress(fromHex("0000000000000000000000000000000000000000000000000000000000000002")),
+      ).toEqual("0x0000000000000000000000000000000000000002");
+      expect(
+        Abi.decodeAddress(fromHex("00000000000000000000000000000000000000000000000000000000075bcd15")),
+      ).toEqual("0x00000000000000000000000000000000075bcd15");
+      expect(
+        Abi.decodeAddress(fromHex("0000000000000000000000003b8a67ad64160e0b977b6dd877e0fb98878ab902")),
+      ).toEqual("0x3b8a67ad64160e0b977b6dd877e0fb98878ab902");
+    });
+
+    it("throws for invalid input", () => {
+      const tooShort = fromHex("00000000000000000000000000000000000000000000000000000000000000");
+      expect(() => Abi.decodeAddress(tooShort)).toThrow();
+
+      const tooLong = fromHex("000000000000000000000000000000000000000000000000000000000000000000");
+      expect(() => Abi.decodeAddress(tooLong)).toThrow();
+    });
+  });
+
+  describe("decodeUint256", () => {
+    it("works", () => {
+      expect(
+        Abi.decodeUint256(fromHex("0000000000000000000000000000000000000000000000000000000000000000")),
+      ).toEqual("0");
+      expect(
+        Abi.decodeUint256(fromHex("0000000000000000000000000000000000000000000000000000000000000001")),
+      ).toEqual("1");
+      expect(
+        Abi.decodeUint256(fromHex("0000000000000000000000000000000000000000000000000000000000000002")),
+      ).toEqual("2");
+      expect(
+        Abi.decodeUint256(fromHex("00000000000000000000000000000000000000000000000000000000075bcd15")),
+      ).toEqual("123456789");
+    });
+
+    it("throws for invalid input", () => {
+      const tooShort = fromHex("00000000000000000000000000000000000000000000000000000000000000");
+      expect(() => Abi.decodeUint256(tooShort)).toThrow();
+
+      const tooLong = fromHex("000000000000000000000000000000000000000000000000000000000000000000");
+      expect(() => Abi.decodeUint256(tooLong)).toThrow();
+    });
+  });
+
   describe("decodeHeadTail", () => {
     it("works for single string", () => {
       const data = fromHex(