+++ scripts/ethereum/init.sh
@@ -6,7 +6,7 @@ export GANACHE_PORT="8545"
 
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-docker pull "alpine"
+# Don't pull before run. We can use whatever version of alpine is available locally.
 DOCKER_HOST_IP=$(docker run --rm alpine ip route | awk 'NR==1 {print $3}')
 
 (