+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## 0.13.1
 
 - @iov/ethereum: Allow querying of ERC20 token balances
+- @iov/ethereum: Allow sending of ERC20 tokens
 - @iov/ethereum: Implement `getFeeQuote` method on `EthereumConnection`.
+- @iov/ethereum: Fix transaction fee parsing for transactions from RPC/scraper.
 - @iov/dpos: Let `Serialization.serializeTransaction` accept transactions with
   fee set.